CSCI-P465/OnlyPets/djangoApi/requirements.txt
asgiref==3.4.1
bidict==0.21.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.46.0
django-cors-headers==3.10.1
django-crispy-forms==1.13.0
django-environ==0.8.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
env==0.1.0
environ==1.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-secret-manager==2.8.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-status==1.42.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
libcst==0.3.23
MarkupSafe==2.0.1
mypy-extensions==0.4.3
oauthlib==3.1.1
Pillow==8.4.0
proto-plus==1.19.8
protobuf==3.19.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
python-engineio==4.3.0
python-socketio==5.5.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
typing-inspect==0.7.1
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
